Change build script from ANT to PHING with parity
Note there is preliminary support for phpdocumentor2 but has been omitted due to some directory bugs.  The intention of this commit is to achieve ANT parity of the final archives in PHING, exclusing YUICompressor.

The build script is being redesigned to be more modular so specific compoents can be built, although not recommeneded at this time.