A11y icons
When reading our default markup for rendering icons, assisistive
technology may have the following problems.

The assistive technology will not find any content to read out to a user
The assistive technology will read the unicode equivalent, which does
not match up to what the icon means in context, or worse is just plain
confusing. In our use case it is always plain wrong. For example the
unicode character used to display the trashed icon is \4c which is equal
to L

When an icon is not an interactive element the simplest way to provide a
text alternative is to use the aria-hidden="true" attribute on the icon
and to include the text with an additional element, such as a < span>,
with appropriate CSS to visually hide the element while keeping it
accessible to assistive technologies.